Add Flask Connection to Application
Flask Application contains model that was pre-trained and saved into
hdf5 files. Essentially, the main function is to generate beam texts
which will get the top 5 responses based on score and return it as an
array.

A basic implementation is done on the Login Page to contain a button
that would make a GET request to the flask application, and retrieve a
smart reply and change the text next to the button.

References:
https://towardsdatascience.com/
training-your-own-message-
suggestions-model-using-deep-learning-3609c0057ba8

https://www.kaggle.com/code/greatrxt/
smartreply-word-level-seq2seq-with-beam-search/notebook

https://towardsdatascience.com/
create-a-complete-machine-learning-web-
application-using-react-and-flask-859340bddb33

https://www.youtube.com/watch?v=7LNl2JlZKHA&t=1s